requirements.txt
discord.py==1.7.3
python-dotenv~=0.19.2